==> code/3. Content Based Recommender System.ipynb <==
# Capstone: Recommender System
## 3. Content Based Recommender System

from nltk import ngrams
import regex as re
import requests
import pandas as pd
import time
import random

import scikitplot as skplt
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity


df = pd.read_csv('../datasets/product_info_clean.csv')

df.drop('Unnamed: 0', axis=1, inplace=True)
df.info()

#create 
df1 = df[['full_name','Category','pure_ingredients']]

#cleaning the ingredients by remove some characters
for i in range(len(df1.pure_ingredients)):
    a = df1.pure_ingredients[i]
    old = ['(',')','  ']
    for char in old:
        a = a.replace(char,' ')
    df1.pure_ingredients[i] = a
    

#create bag-of-word
#column for unigram and column for combitation of unigram and bigram
df1.loc[:,'unigram'] = np.nan
df1.loc[:,'uni_bigram'] = np.nan

df1.head()

#method 1
#unigram
for i in range(len(df1.pure_ingredients)):
    ingredients_tokens = tokenizer.tokenize(df1.pure_ingredients[i].lower())

    df1.unigram[i] = ingredients_tokens
    

#method 2
#combination of uigram and bigram

for i in range(len(df1.pure_ingredients)):
    ingredients_tokens = df1.pure_ingredients[i].lower().split(', ')
    l = []
    for token in ingredients_tokens:
        if ' ' in token:
            bigrams = ngrams(token.split(), 2)
            l.extend(bigrams)
        else:
            l.append(token)
    df1.uni_bigram[i] = l
    

df1.head()

#read in chemicals file
#chemicals file consists various chemicals that give products different effects
df2 = pd.read_csv('../datasets/chemicals.csv', encoding='latin1')
df2.feature.value_counts()

#create feature list
feature_list = ['moisturising effect','cleanliness','antioxidant','whitening','soothing','irritancy','viscosity','fragrance']
df_feature = pd.DataFrame({'feature':feature_list,'chemicals_uni':'','chemicals_unibi':''}) 

#method 1 unigram
for i in range(len(feature_list)):
    chem = df2[df2.feature == feature_list[i]].chemicals
    l = []
    for c in chem:
        l=l+tokenizer.tokenize(c.lower())

        df_feature.chemicals_uni[i] = l
    
    

#method 2 combination of unigram and bigram
for i in range(len(feature_list)):
    chem = df2[df2.feature == feature_list[i]].chemicals
    l = []
    for c in chem:
        if ' ' in c:
            bigrams = ngrams(c.lower().split(), 2)
            l.extend(bigrams)
        else:
            l.append(c.lower())
    df_feature.chemicals_unibi[i] = l



df_feature

#jaccard_similarity
def jaccard_similarity(list1, list2):
    s1 = set(list1)
    s2 = set(list2)
    return len(s1.intersection(s2)) / len(s1.union(s2))


#using different method for different features
feature_list_1 = ['moisturising effect','whitening','viscosity']
feature_list_2 = ['cleanliness','antioxidant','fragrance']
def ingre_similarity(product, feature, chemical):
    for e in feature:
        s = []
        length = len(product)
        n = df_feature[df_feature['feature'] == e].index.values[0]
        for i in range(length):
            
            score1 = jaccard_similarity(product[i][0:int(length/3)], chemical[n])  
            score2 = jaccard_similarity(product[i][int(length/3):int(length*2/3)], chemical[n])
            score3 = jaccard_similarity(product[i][int(length*2/3):length], chemical[n])
            score = 0.5*score1 + 0.3*score2 +0.2*score3   
            s.append(score)
        df1[e] = s
        
ingre_similarity(df1.unigram, feature_list_1, df_feature.chemicals_uni)
ingre_similarity(df1.uni_bigram, feature_list_2, df_feature.chemicals_unibi)

df1.sort_values(by='moisturising effect', ascending=False)[['full_name','Category','moisturising effect']].head()

df1.sort_values(by='cleanliness', ascending=False)[['full_name','Category','cleanliness']].head()

df1.sort_values(by='viscosity', ascending=False)[['full_name','Category','viscosity']].head()


df1.sort_values(by='whitening', ascending=False)[['full_name','Category','whitening']].head()

df1.sort_values(by='antioxidant', ascending=False)[['full_name','Category','antioxidant']].head(10)

df1.sort_values(by='fragrance', ascending=False)[['full_name','Category','fragrance']].head()

##part 2 sentiment analysis for soothing and irritancy 
#in this case, we will treat soothing as positive and irritancy as negetive
#use combination of unigram and bigram
def simple_sentiment(ingredient):
    
    # positive/negative 
    index1 = df_feature[df_feature['feature'] == 'soothing'].index.values[0]
    index2 = df_feature[df_feature['feature'] == 'irritancy'].index.values[0]
    positive = df_feature.chemicals_unibi[index1]
    negative = df_feature.chemicals_unibi[index2]

    # Count 
    positive_count = sum([1 for i in ingredient if i in positive])
    negative_count = sum([1 for i in ingredient if i in negative])
    
    # Calculate Sentiment Percentage 
    # (Positive Count - Negative Count) / (Total Count)

    return round((positive_count - negative_count) / len(ingredient), 6)

##mildness score 　
l = []
for ingre in df1.uni_bigram:
    score = simple_sentiment(ingre)
    l.append(score)
df1.loc[:,'mildness'] = l 

df1[['full_name','mildness']].sort_values(by = 'mildness',ascending = False).head()

df1.head()

#adding product information
product_score1 = df[['full_name','brand', 'Category','price_oz']]
product_score2 = df1[['moisturising effect','cleanliness','mildness','antioxidant','whitening','viscosity','fragrance']]
product_score = pd.concat([product_score1, product_score2], axis = 1)



product_score.head()

##normalize tha score to the range of 0-1
def norm(series):
    l = []
    for i in range(len(series)):
        n = (series[i] - series.min()) / (series.max() - series.min())
        l.append(n)
    series = l
        

features = ['price_oz','moisturising effect','cleanliness','mildness','antioxidant','whitening','viscosity','fragrance']
for feature in features:
    series = product_score[feature]
    l = []
    for i in range(len(series)):
        n = (series[i] - series.min()) / (series.max() - series.min())
        l.append(n)
    product_score[feature] = l

product_score.to_csv('../datasets/product_score.csv')

## Cosine similarity

product_score = pd.read_csv('../datasets/product_score.csv')

product_score.drop(columns = 'Unnamed: 0', axis =1, inplace = True)

content = product_score.iloc[:,3:11]

content.head()

#Cosine similarity
sim_matrix = cosine_similarity(content)
content_sim = pd.DataFrame(sim_matrix, columns=content.index, index=content.index)


content_sim.to_csv('../datasets/content_sim.csv')
