==> code/4. Collaborative Recommender System.ipynb <==
# Capstone: Recommender System
## 4. Collaborative Recommender System

import pandas as pd
import numpy as np
from surprise import SVD, KNNWithMeans, SVDpp, KNNBasic
from surprise import accuracy
from surprise import Reader, Dataset
from surprise.model_selection import train_test_split
from surprise.model_selection import GridSearchCV
from surprise.model_selection import cross_validate
from scipy.sparse.linalg import svds
from sklearn.metrics.pairwise import cosine_similarity

pip install surprise

review = pd.read_csv('../datasets/review_clean.csv')

review.info()

review.drop(columns = ['Unnamed: 0'], inplace = True)

#constract a users vs product matrix, it contains rating products from users
up = pd.pivot_table(review, index='user', columns='id', values='stars').fillna(0)

up

#find users who only purchase one product
l = []
for i in up.index:
    s = sum([1 for i in up.loc[i] if i != 0.0])
    if s < 2:
        l.append(i)

len(l)

#drop
up.drop(index = l, inplace = True)

len(up)

#find one time purchased users index number in review dataset
indexlist = []
for i in l:
    n = review[review.user == i].index[0]
    indexlist.append(n)

#drop those users for review dataframe and create a new dataset review1
review1 = review[['user', 'id', 'stars']]
review1 = review1.drop(index = indexlist)


len(review1)

## Model Evaluation with package surprise

#data preparetions
# set rating_scale param 
reader = Reader(rating_scale=(1, 5))

# The columns must correspond to user id, item id and ratings (in that order).
data = Dataset.load_from_df(review1[['user', 'id', 'stars']], reader)

#train test split
trainset, testset = train_test_split(data, test_size = 0.25)



## SVD

#Grid Search for the best parameters
param_grid = {'n_factors':[40, 60], 
              'n_epochs': [20, 40, 60],
              'lr_all': [0.002, 0.005],
              'reg_all': [0.1, 0.4, 0.6]}
gs = GridSearchCV(SVD, param_grid, measures=['rmse', 'mae'], cv=3)

gs.fit(data)
# best RMSE score
print(gs.best_score['rmse'])
# combination of parameters that gave the best RMSE score
print(gs.best_params['rmse'])

#run the SVD model
algo = gs.best_estimator['rmse']
algo.fit(trainset)
test_pred = algo.test(testset)
accuracy.rmse(test_pred)

## SVD++

#Grid Search for the best parameters
param_grid = {'n_factors':[40, 60], 'n_epochs': [20, 40], 'lr_all': [0.005, 0.007],
              'reg_all': [0.2, 0.4, 0.6]}
gs1 = GridSearchCV(SVDpp, param_grid, measures=['rmse', 'mae'], cv=3)

gs1.fit(data)

# best RMSE score
print(gs1.best_score['rmse'])

# combination of parameters that gave the best RMSE score
print(gs1.best_params['rmse'])

algo = gs1.best_estimator['rmse']
algo.fit(trainset)
test_pred = algo.test(testset)
accuracy.rmse(test_pred)

## KNN

algo = KNNBasic(k=40, sim_options={'name': 'pearson_baseline', 'user_based': False})
algo.fit(trainset)

# run the trained model against the testset
test_pred = algo.test(testset)

# get RMSE
accuracy.rmse(test_pred, verbose=True)

By comparing the RMSE values of three models, SVD and SVD++ give us better result.

And then, we will construct a matix with SVD.


up_matrix = up.values

#Performs matrix factorization of the original user item matrix
U, sigma, Vt = svds(up_matrix , k = 40)

sigma = np.diag(sigma)
sigma.shape

After the factorization, we try to to reconstruct the original matrix by multiplying its factors. The resulting matrix is not sparse any more. It was generated predictions for items the user have not yet interaction, which we will exploit for recommendations.

all_user_predicted_ratings = np.dot(np.dot(U, sigma), Vt) 
all_user_predicted_ratings.shape

#Converting the reconstructed matrix back to a Pandas dataframe
item_based = pd.DataFrame(all_user_predicted_ratings, columns = up.columns, index=list(up.index)).transpose()
item_based.shape


sim_matrix = cosine_similarity(item_based)
collaborative_sim = pd.DataFrame(sim_matrix, columns=item_based.index, index=item_based.index)

collaborative_sim.head()

collaborative_sim.to_csv('../datasets/collaborative_sim.csv')
