==> code/5. Aggregation and Recommendation.ipynb <==
# Capstone: Recommender System
## 5. Aggregation and Recommendation

import pandas as pd
from html.parser import HTMLParser
import seaborn as sns
import numpy as np

df = pd.read_csv('../datasets/product_info_clean.csv')
content = pd.read_csv('../datasets/content_sim.csv')
collab = pd.read_csv('../datasets/collaborative_sim.csv')
repurchase = pd.read_csv('../datasets/repurchase.csv')

#drop Unnamed: 0 columns
df.drop(columns = 'Unnamed: 0', axis = 1, inplace = True)
collab.drop(columns = 'id', axis =1 , inplace =True)
content.drop(columns = 'Unnamed: 0', axis = 1, inplace = True)
repurchase.drop(columns = 'Unnamed: 0', axis = 1, inplace = True)


df.Category.value_counts()

collab

id 0 - 92 are facial cleansers, id 93 - 185 are toners and id 186 - 280 are moisturizers.

Define the recommender function, the inputs are the category of products users are looking for and the name of product they are using right now.

def recommender(cat, product_name):
    product_id = df[df.full_name == product_name].index[0]
    
    #index of products from desired category
    cat_in = df[df.Category == cat].index
        
    #rank product by similarity, lower the ranking number, higher the similarity
    content_rank = content.iloc[product_id,cat_in].rank(ascending=False, method='min')
    collab_rank = collab.iloc[product_id,cat_in].rank(ascending=False, method='min')
    
    #aggregate two rankings to get overall rankings
    rank = content_rank + collab_rank
    
    #sort rankings with ascending orders, products on top will most likely be recommended
    rank = rank.sort_values()[1:]
    
    #for products have same rankings after aggragation, we will take into account the repurchase rate.
    re_one = rank[rank.duplicated()]
    re_all = rank[rank.duplicated(keep=False)]
    
    for i in re_one:
        list_index = []
        list_num = []
        for e in re_all.index:
            if i == re_all[e]:
                list_index.append(e)
                list_num.append(int(e))
        r = repurchase.loc[list_num,'rate'].rank(ascending = False,method = 'min')
        rank[list_index] = rank[list_index] + list(0.01*r)   
            
    #show the top 10 products' infomation  
    result = df.iloc[list(rank.sort_values()[:10].index)]
    
    #show cosine similarity score for both content base and collaborative
    content_sim = list(content.iloc[product_id,list(result.index)])
    collab_sim = list(collab.iloc[product_id,list(result.index)])
    result.loc[:,'content_sim'] = content_sim
    result.loc[:,'collab_sim'] = collab_sim
    final = result[['full_name','brand','content_sim','collab_sim']]
    return final
